* Lock viewer placed in separate plugin org.jkiss.dbeaver.ext.ui.locks in pursuance of Serge review :
Could you extract org.jkiss.dbeaver.ui.views.lock* into separate plugin? Let's say org.jkiss.dbeaver.ext.ui.locks. This will make you a plugin owner,
will simplify its support and will remove gef/draw2d dependency from core module.

All reference to GEF removed from core.

-----------

+ Also some minor fix.
