--- plugins/org.jkiss.dbeaver.core/src/org/jkiss/dbeaver/ui/views/lock/LockGraphConnectionEditPolicy.java
@@ -1,30 +1,0 @@
-/*
- * DBeaver - Universal Database Manager
- * Copyright (C) 2017 Andrew Khitrin (ahitrin@gmail.com) 
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.jkiss.dbeaver.ui.views.lock;
-
-import org.eclipse.gef.commands.Command;
-import org.eclipse.gef.editpolicies.ConnectionEditPolicy;
-import org.eclipse.gef.requests.GroupRequest;
-
-public class LockGraphConnectionEditPolicy extends ConnectionEditPolicy {
-
-	@Override
-	protected Command getDeleteCommand(GroupRequest request) {
-		return null;
-	}
-
-}
