+++ plugins/org.jkiss.dbeaver.ext.oracle/META-INF/MANIFEST.MF
@@ -21,3 +21,5 @@ Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Bundle-Vendor: JKISS
 Bundle-ClassPath: .
 Bundle-Localization: plugin
+Import-Package: org.jkiss.dbeaver.ext.ui.locks.edit,
+ org.jkiss.dbeaver.ext.ui.locks.manage
