+++ plugins/org.jkiss.dbeaver.ext.postgresql/META-INF/MANIFEST.MF
@@ -21,3 +21,5 @@ Bundle-Vendor: JKISS
 Bundle-ClassPath: .
 Bundle-Localization: plugin
 Bundle-Activator: org.jkiss.dbeaver.ext.postgresql.PostgreActivator
+Import-Package: org.jkiss.dbeaver.ext.ui.locks.edit,
+ org.jkiss.dbeaver.ext.ui.locks.manage
