+++ plugins/org.jkiss.dbeaver.ext.postgresql/src/org/jkiss/dbeaver/ext/postgresql/edit/PostgreLockEditor.java
@@ -25,12 +25,12 @@ import org.eclipse.swt.widgets.Composite;
 import org.jkiss.dbeaver.ext.postgresql.model.PostgreDataSource;
 import org.jkiss.dbeaver.ext.postgresql.model.lock.PostgreLock;
 import org.jkiss.dbeaver.ext.postgresql.model.lock.PostgreLockManager;
+import org.jkiss.dbeaver.ext.ui.locks.edit.AbstractLockEditor;
+import org.jkiss.dbeaver.ext.ui.locks.manage.LockManagerViewer;
 import org.jkiss.dbeaver.model.admin.locks.DBAServerLock;
 import org.jkiss.dbeaver.model.admin.locks.DBAServerLockItem;
 import org.jkiss.dbeaver.model.admin.locks.DBAServerLockManager;
 import org.jkiss.dbeaver.model.exec.DBCExecutionContext;
-import org.jkiss.dbeaver.ui.views.lock.AbstractLockEditor;
-import org.jkiss.dbeaver.ui.views.lock.LockManagerViewer;
 
 public class PostgreLockEditor extends AbstractLockEditor{
